+++ R/package.R
@@ -18,7 +18,7 @@ renv_package_find_impl <- function(package,
 {
   # if we've been given the path to an existing package, use it as-is
   if (file.exists(file.path(package, "DESCRIPTION")))
-    return(normalizePath(package, winslash = "/", mustWork = TRUE))
+    return(renv_path_normalize(package, mustWork = TRUE))
 
   # first, look in the library paths
   for (libpath in lib.loc) {
@@ -276,7 +276,7 @@ renv_package_dependencies_impl <- function(package,
   assign(package, location, envir = visited, inherits = FALSE)
 
   # find its dependencies from the DESCRIPTION file
-  deps <- renv_dependencies_discover_description(location, fields)
+  deps <- renv_dependencies_discover_description(location, fields = "strong")
   subpackages <- deps$Package
   for (subpackage in subpackages)
     renv_package_dependencies_impl(subpackage, visited, libpaths, fields)