+++ R/project.R
@@ -43,7 +43,7 @@ renv_project_clear <- function() {
 
 renv_project_resolve <- function(project = NULL, default = getwd()) {
   project <- project %||% renv_project_get(default = default)
-  normalizePath(project, winslash = "/", mustWork = FALSE)
+  renv_path_normalize(project)
 }
 
 renv_project_initialized <- function(project) {
@@ -101,35 +101,16 @@ renv_project_type_impl <- function(path) {
 
 renv_project_remotes <- function(project, fields = NULL) {
 
-  # if this project has a DESCRIPTION file, use it to provide records
   descpath <- file.path(project, "DESCRIPTION")
-  if (file.exists(descpath))
-    return(renv_project_remotes_description(project, descpath, fields))
-
-  # otherwise, use the set of (non-base) packages used in the project
-  deps <- dependencies(
-    path     = project,
-    progress = FALSE,
-    errors   = "ignored",
-    dev      = TRUE
-  )
-
-  packages <- sort(unique(deps$Package))
-  ignored <- c("renv", renv_project_ignored_packages(project))
-  setdiff(packages, ignored)
-
-}
-
-renv_project_remotes_description <- function(project, descpath, fields = NULL) {
+  if (!file.exists(descpath))
+    return(NULL)
 
   # first, parse remotes (if any)
-  remotes <- renv_project_remotes_description_remotes(project, descpath)
+  remotes <- renv_project_remotes_field(project, descpath)
 
   # next, find packages mentioned in the DESCRIPTION file
-  fields <- fields %||% c("Depends", "Imports", "LinkingTo", "Suggests")
   deps <- renv_dependencies_discover_description(
     path    = descpath,
-    fields  = fields,
     project = project
   )
 
@@ -182,7 +163,7 @@ renv_project_remotes_description <- function(project, descpath, fields = NULL) {
 
 }
 
-renv_project_remotes_description_remotes <- function(project, descpath) {
+renv_project_remotes_field <- function(project, descpath) {
 
   desc <- renv_description_read(descpath)