+++ tests/testthat/test-paths.R
@@ -55,7 +55,7 @@ test_that("UTF-8 paths can be normalized", {
     skip("locale is not UTF-8")
 
   name <- enc2utf8("\u4f60\u597d")  # nihao
-  root <- normalizePath(tempdir(), winslash = "/", mustWork = TRUE)
+  root <- renv_path_normalize(tempdir())
   path <- paste(root, name, sep = "/")
   defer(unlink(path))
   expect_true(file.create(path))